requirements.txt
aiofiles
aiohttp
asyncio
beautifulsoup4
dnspython
ffmpeg-python
gitpython
hachoir
heroku3
motor
pillow==10.2.0
psutil
py-tgcalls==0.9.2
pykeyboard
pyrogram==2.0.106
python-dotenv
pyyaml
requests
speedtest-cli
spotipy
tgcrypto
unidecode
yt-dlp
youtube-search
youtube-search-python